refactor: relocate auth infra/security implementations out of domain
