+++ src/main/java/com/ticketrush/api/controller/AuthController.java
@@ -4,10 +4,10 @@ import com.ticketrush.api.dto.auth.AuthMeResponse;
 import com.ticketrush.api.dto.auth.AuthTokenResponse;
 import com.ticketrush.api.dto.auth.TokenLogoutRequest;
 import com.ticketrush.api.dto.auth.TokenRefreshRequest;
+import com.ticketrush.application.auth.model.AuthUserPrincipal;
 import com.ticketrush.application.auth.service.AuthSessionService;
 import com.ticketrush.application.user.service.UserService;
 import com.ticketrush.domain.auth.model.AuthTokenPair;
-import com.ticketrush.domain.auth.security.AuthUserPrincipal;
 import com.ticketrush.domain.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
