+++ src/main/java/com/ticketrush/application/auth/service/AuthSessionServiceImpl.java
@@ -4,7 +4,6 @@ import com.ticketrush.domain.auth.entity.RefreshToken;
 import com.ticketrush.domain.auth.model.AuthTokenPair;
 import com.ticketrush.domain.auth.repository.RefreshTokenRepository;
 import com.ticketrush.domain.auth.service.AccessTokenDenylistService;
-import com.ticketrush.domain.auth.service.JwtTokenProvider;
 import com.ticketrush.domain.user.User;
 import com.ticketrush.domain.user.UserRepository;
 import io.jsonwebtoken.Claims;
