+++ src/test/java/com/ticketrush/api/controller/AuthSecurityIntegrationTest.java
@@ -1,10 +1,10 @@
 package com.ticketrush.api.controller;
 
 import com.ticketrush.application.auth.service.AuthSessionService;
+import com.ticketrush.application.auth.service.JwtTokenProvider;
 import com.ticketrush.application.user.service.UserService;
-import com.ticketrush.domain.auth.security.JwtAuthenticationFilter;
 import com.ticketrush.domain.auth.service.AccessTokenDenylistService;
-import com.ticketrush.domain.auth.service.JwtTokenProvider;
+import com.ticketrush.infrastructure.auth.security.JwtAuthenticationFilter;
 import com.ticketrush.domain.user.User;
 import com.ticketrush.domain.user.UserRole;
 import com.ticketrush.domain.user.UserTier;
