+++ src/test/java/com/ticketrush/api/controller/SocialAuthControllerIntegrationTest.java
@@ -1,13 +1,13 @@
 package com.ticketrush.api.controller;
 
 import com.ticketrush.application.auth.service.AuthSessionService;
+import com.ticketrush.application.auth.service.JwtTokenProvider;
 import com.ticketrush.application.auth.service.SocialAuthService;
 import com.ticketrush.domain.auth.model.AuthTokenPair;
 import com.ticketrush.domain.auth.model.SocialAuthorizeInfo;
 import com.ticketrush.domain.auth.model.SocialLoginResult;
-import com.ticketrush.domain.auth.security.JwtAuthenticationFilter;
 import com.ticketrush.domain.auth.service.AccessTokenDenylistService;
-import com.ticketrush.domain.auth.service.JwtTokenProvider;
+import com.ticketrush.infrastructure.auth.security.JwtAuthenticationFilter;
 import com.ticketrush.domain.user.SocialProvider;
 import com.ticketrush.domain.user.User;
 import com.ticketrush.domain.user.UserTier;
