+++ src/test/java/com/ticketrush/application/auth/service/AuthSessionServiceTest.java
@@ -4,8 +4,7 @@ import com.ticketrush.domain.auth.entity.RefreshToken;
 import com.ticketrush.domain.auth.model.AuthTokenPair;
 import com.ticketrush.domain.auth.repository.RefreshTokenRepository;
 import com.ticketrush.domain.auth.service.AccessTokenDenylistService;
-import com.ticketrush.domain.auth.service.InMemoryAccessTokenDenylistService;
-import com.ticketrush.domain.auth.service.JwtTokenProvider;
+import com.ticketrush.infrastructure.auth.denylist.InMemoryAccessTokenDenylistService;
 import com.ticketrush.domain.user.User;
 import com.ticketrush.domain.user.UserRepository;
 import com.ticketrush.domain.user.UserRole;
