+++ src/test/java/com/ticketrush/architecture/LayerDependencyArchTest.java
@@ -24,6 +24,12 @@ class LayerDependencyArchTest {
                     .should().dependOnClassesThat().resideInAnyPackage("com.ticketrush.application..");
 
     @ArchTest
+    static final ArchRule domain_should_not_depend_on_infrastructure_layer =
+            noClasses()
+                    .that().resideInAPackage("com.ticketrush.domain..")
+                    .should().dependOnClassesThat().resideInAnyPackage("com.ticketrush.infrastructure..");
+
+    @ArchTest
     static final ArchRule application_should_not_depend_on_api_layer =
             noClasses()
                     .that().resideInAPackage("com.ticketrush.application..")
